+++ app/upgrades/v2_2/upgrade.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"fmt"
 
+	liquidkeeper "github.com/KYVENetwork/chain/x/liquid/keeper"
+	stakingKeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
+
 	"cosmossdk.io/log"
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 
@@ -20,6 +23,8 @@ var logger log.Logger
 func CreateUpgradeHandler(
 	mm *module.Manager,
 	configurator module.Configurator,
+	liquidStakingKeeper *liquidkeeper.Keeper,
+	stakingKeeper *stakingKeeper.Keeper,
 ) upgradetypes.UpgradeHandler {
 	return func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 		sdkCtx := sdk.UnwrapSDKContext(ctx)
@@ -30,6 +35,14 @@ func CreateUpgradeHandler(
 		migratedVersionMap, err := mm.RunMigrations(ctx, configurator, fromVM)
 
 		// Run KYVE migrations
+		validators, _ := stakingKeeper.GetValidators(ctx, 1_000)
+		for _, v := range validators {
+			valAddress, _ := sdk.ValAddressFromBech32(v.OperatorAddress)
+			err := liquidStakingKeeper.Hooks().AfterValidatorCreated(ctx, valAddress)
+			if err != nil {
+				return nil, err
+			}
+		}
 
 		logger.Info(fmt.Sprintf("finished upgrade %v", UpgradeName))